5.48

+++ src/main/java/jp/co/toshiba/ppok/repository/CityDao.java
@@ -1,12 +1,13 @@
 package jp.co.toshiba.ppok.repository;
 
-import jp.co.toshiba.ppok.entity.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
+
+import jp.co.toshiba.ppok.entity.City;
 
 /**
  * searching dao of table WORLD_CITY
@@ -15,15 +16,15 @@ import org.springframework.transaction.annotation.Transactional;
  * @date 2022-12-17
  */
 @Repository
-public interface CityDao extends JpaRepository<City, Long> {
+public interface CityDao extends JpaRepository<City, Integer> {
 
-    /**
-     * logic remove query.
-     *
-     * @param id id of the selected city
-     */
-    @Modifying
-    @Transactional
-    @Query(value = "update City cvn set cvn.isDeleted = 1 where cvn.id =:id")
-    void removeById(@Param("id") final Long id);
+	/**
+	 * logic remove query.
+	 *
+	 * @param id id of the selected city
+	 */
+	@Modifying
+	@Transactional
+	@Query(value = "update City cvn set cvn.isDeleted = 1 where cvn.id =:id")
+	void removeById(@Param("id") final Integer id);
 }
